ksn_server: refactor key_share into v1/v2 files
