+++ key_share_node/server/src/routes/key_share/v2.ts
@@ -1,0 +1,317 @@
+import { Router, type Response } from "express";
+import type {
+  CheckKeyShareV2Request,
+  CheckKeyShareV2RequestBody,
+  CheckKeyShareV2Response,
+  GetKeyShareV2Request,
+  GetKeyShareV2RequestBody,
+  GetKeyShareV2Response,
+} from "@oko-wallet/ksn-interface/key_share";
+import { Bytes } from "@oko-wallet/bytes";
+import type { KSNodeApiResponse } from "@oko-wallet/ksn-interface/response";
+
+import {
+  checkKeyShareV2,
+  getKeyShareV2,
+} from "@oko-wallet-ksn-server/api/key_share";
+import {
+  bearerTokenMiddleware,
+  type AuthenticatedRequest,
+} from "@oko-wallet-ksn-server/middlewares";
+import { ErrorCodeMap } from "@oko-wallet-ksn-server/error";
+import type {
+  ResponseLocal,
+  KSNodeRequest,
+} from "@oko-wallet-ksn-server/routes/io";
+import { registry } from "@oko-wallet-ksn-server/openapi/registry";
+import {
+  CheckKeyShareV2RequestBodySchema,
+  CheckKeyShareV2SuccessResponseSchema,
+  GetKeyShareV2RequestBodySchema,
+  GetKeyShareV2SuccessResponseSchema,
+  ErrorResponseSchema,
+} from "@oko-wallet-ksn-server/openapi/schema";
+
+export function makeKeyshareV2Router() {
+  const router = Router();
+
+  registry.registerPath({
+    method: "post",
+    path: "/keyshare/v2",
+    tags: ["Key Share v2"],
+    summary: "Get multiple key shares",
+    description:
+      "Retrieve multiple key shares for the authenticated user in a single request.",
+    security: [{ oauthAuth: [] }],
+    request: {
+      body: {
+        required: true,
+        content: {
+          "application/json": {
+            schema: GetKeyShareV2RequestBodySchema,
+          },
+        },
+      },
+    },
+    responses: {
+      200: {
+        description: "Successfully retrieved key shares",
+        content: {
+          "application/json": {
+            schema: GetKeyShareV2SuccessResponseSchema,
+          },
+        },
+      },
+      400: {
+        description: "Bad request",
+        content: {
+          "application/json": {
+            schema: ErrorResponseSchema,
+            examples: {
+              PUBLIC_KEY_INVALID: {
+                value: {
+                  success: false,
+                  code: "PUBLIC_KEY_INVALID",
+                  msg: "Public key is not valid",
+                },
+              },
+            },
+          },
+        },
+      },
+      401: {
+        description: "Unauthorized - Invalid or missing bearer token",
+        content: {
+          "application/json": {
+            schema: ErrorResponseSchema,
+          },
+        },
+      },
+      404: {
+        description: "Not found - User, wallet or key share not found",
+        content: {
+          "application/json": {
+            schema: ErrorResponseSchema,
+            examples: {
+              USER_NOT_FOUND: {
+                value: {
+                  success: false,
+                  code: "USER_NOT_FOUND",
+                  msg: "User not found",
+                },
+              },
+              WALLET_NOT_FOUND: {
+                value: {
+                  success: false,
+                  code: "WALLET_NOT_FOUND",
+                  msg: "Wallet not found for curve_type: secp256k1",
+                },
+              },
+              KEY_SHARE_NOT_FOUND: {
+                value: {
+                  success: false,
+                  code: "KEY_SHARE_NOT_FOUND",
+                  msg: "Key share not found for curve_type: ed25519",
+                },
+              },
+            },
+          },
+        },
+      },
+      500: {
+        description: "Internal server error",
+        content: {
+          "application/json": {
+            schema: ErrorResponseSchema,
+          },
+        },
+      },
+    },
+  });
+  router.post(
+    "/",
+    bearerTokenMiddleware,
+    async (
+      req: AuthenticatedRequest<GetKeyShareV2RequestBody>,
+      res: Response<KSNodeApiResponse<GetKeyShareV2Response>, ResponseLocal>,
+    ) => {
+      const oauthUser = res.locals.oauth_user;
+      const auth_type = oauthUser.type;
+      const state = req.app.locals;
+      const body = req.body;
+
+      // Validate and convert wallets object
+      const validatedWallets: GetKeyShareV2Request["wallets"] = {};
+
+      // Validate secp256k1
+      if (body.wallets.secp256k1) {
+        const publicKeyBytesRes = Bytes.fromHexString(
+          body.wallets.secp256k1,
+          33,
+        );
+        if (publicKeyBytesRes.success === false) {
+          return res.status(400).json({
+            success: false,
+            code: "PUBLIC_KEY_INVALID",
+            msg: `Public key is not valid for secp256k1: ${publicKeyBytesRes.err}`,
+          });
+        }
+        validatedWallets.secp256k1 = publicKeyBytesRes.data;
+      }
+
+      // Validate ed25519
+      if (body.wallets.ed25519) {
+        const publicKeyBytesRes = Bytes.fromHexString(body.wallets.ed25519, 32);
+        if (publicKeyBytesRes.success === false) {
+          return res.status(400).json({
+            success: false,
+            code: "PUBLIC_KEY_INVALID",
+            msg: `Public key is not valid for ed25519: ${publicKeyBytesRes.err}`,
+          });
+        }
+        validatedWallets.ed25519 = publicKeyBytesRes.data;
+      }
+
+      const result = await getKeyShareV2(
+        state.db,
+        {
+          user_auth_id: oauthUser.user_identifier,
+          auth_type,
+          wallets: validatedWallets,
+        },
+        state.encryptionSecret,
+      );
+
+      if (result.success === false) {
+        return res.status(ErrorCodeMap[result.code]).json({
+          success: false,
+          code: result.code,
+          msg: result.msg,
+        });
+      }
+
+      return res.status(200).json({
+        success: true,
+        data: result.data,
+      });
+    },
+  );
+
+  // --- POST /check ---
+  registry.registerPath({
+    method: "post",
+    path: "/keyshare/v2/check",
+    tags: ["Key Share v2"],
+    summary: "Check multiple key shares existence",
+    description:
+      "Check existence of multiple key shares for a user in a single request. No authentication required.",
+    request: {
+      body: {
+        required: true,
+        content: {
+          "application/json": {
+            schema: CheckKeyShareV2RequestBodySchema,
+          },
+        },
+      },
+    },
+    responses: {
+      200: {
+        description: "Successfully checked key shares existence",
+        content: {
+          "application/json": {
+            schema: CheckKeyShareV2SuccessResponseSchema,
+          },
+        },
+      },
+      400: {
+        description: "Bad request",
+        content: {
+          "application/json": {
+            schema: ErrorResponseSchema,
+            examples: {
+              PUBLIC_KEY_INVALID: {
+                value: {
+                  success: false,
+                  code: "PUBLIC_KEY_INVALID",
+                  msg: "Public key is not valid",
+                },
+              },
+            },
+          },
+        },
+      },
+      500: {
+        description: "Internal server error",
+        content: {
+          "application/json": {
+            schema: ErrorResponseSchema,
+          },
+        },
+      },
+    },
+  });
+  router.post(
+    "/check",
+    async (
+      req: KSNodeRequest<CheckKeyShareV2RequestBody>,
+      res: Response<KSNodeApiResponse<CheckKeyShareV2Response>, ResponseLocal>,
+    ) => {
+      const state = req.app.locals;
+      const body = req.body;
+
+      // Validate and convert wallets object
+      const validatedWallets: CheckKeyShareV2Request["wallets"] = {};
+
+      // Validate secp256k1
+      if (body.wallets.secp256k1) {
+        const publicKeyBytesRes = Bytes.fromHexString(
+          body.wallets.secp256k1,
+          33,
+        );
+        if (publicKeyBytesRes.success === false) {
+          return res.status(400).json({
+            success: false,
+            code: "PUBLIC_KEY_INVALID",
+            msg: `Public key is not valid for secp256k1: ${publicKeyBytesRes.err}`,
+          });
+        }
+        validatedWallets.secp256k1 = publicKeyBytesRes.data;
+      }
+
+      // Validate ed25519
+      if (body.wallets.ed25519) {
+        const publicKeyBytesRes = Bytes.fromHexString(body.wallets.ed25519, 32);
+        if (publicKeyBytesRes.success === false) {
+          return res.status(400).json({
+            success: false,
+            code: "PUBLIC_KEY_INVALID",
+            msg: `Public key is not valid for ed25519: ${publicKeyBytesRes.err}`,
+          });
+        }
+        validatedWallets.ed25519 = publicKeyBytesRes.data;
+      }
+
+      const result = await checkKeyShareV2(state.db, {
+        user_auth_id: body.user_auth_id,
+        auth_type: body.auth_type,
+        wallets: validatedWallets,
+      });
+
+      if (result.success === false) {
+        return res.status(ErrorCodeMap[result.code]).json({
+          success: false,
+          code: result.code,
+          msg: result.msg,
+        });
+      }
+
+      return res.status(200).json({
+        success: true,
+        data: result.data,
+      });
+    },
+  );
+
+  return router;
+}
